feat(gastby-theme-jamfire-conference): extracted event context from main theme context
