+++ packages/gatsby-theme-jamfire-conference/src/services/event/event-context.tsx
@@ -1,0 +1,17 @@
+// import libs
+import { createContext } from "react"
+import { InitialState } from "./event.d"
+
+// initial state
+export const initialState: InitialState = {
+  activeRoom: false,
+  setActiveRoom: () => null,
+  roomModal: false,
+  toggleRoomModal: () => true || false,
+  geolocation: null,
+  setGeolocation: () => null,
+}
+
+export const EventContext = createContext(initialState)
+
+export default EventContext
