+++ packages/gatsby-theme-jamfire-conference/src/services/event/index.tsx
@@ -1,0 +1,4 @@
+import EventContext, { initialState } from "./event-context"
+import EventProvider from "./event-provider"
+
+export { EventContext, EventProvider, initialState }
